refactor: restructure methods
