+++ my-first-react-app/src/view/results/Results.tsx
@@ -6,34 +6,37 @@ import Ship from './Ship';
 class Results extends Component {
   declare props: { searchValue: string };
 
-  state: Readonly<{ results: ResultsData[] }> = {
-    results: [{ name: '', model: '', length: '' }],
+  declare state: Readonly<{ results: ResultsData[] }>;
+
+  showContent = (st: { results: ResultsData[] }) => {
+    if (st) {
+      return st.results.map((res, i) => {
+        return <Ship key={i} name={res.name} model={res.model} length={res.length} />;
+      });
+    } else {
+      return '';
+    }
   };
 
   render() {
-    return (
-      <section className="results">
-        {this.state.results.map((res, i) => {
-          return <Ship key={i} name={res.name} model={res.model} length={res.length} />;
-        })}
-      </section>
-    );
+    return <section className="results">{this.showContent(this.state)}</section>;
   }
-  async componentDidMount(): Promise<void> {
+
+  getData = async () => {
     const searchData = await searchRequest(this.props.searchValue);
     const results = searchData.results;
     this.setState({
       results: results,
     });
+  };
+
+  async componentDidMount(): Promise<void> {
+    await this.getData();
   }
 
   async componentDidUpdate(prevProps: Readonly<{ searchValue: string }>): Promise<void> {
     if (prevProps.searchValue !== this.props.searchValue) {
-      const searchData = await searchRequest(this.props.searchValue);
-      const results = searchData.results;
-      this.setState({
-        results: results,
-      });
+      await this.getData();
     }
   }
 }
